+++ src/main/java/com/library/service/CopyService.java
@@ -13,24 +13,25 @@ import java.util.List;
 public class CopyService {
     private final CopyRepository copyRepository;
 
-    public List<Copy> getCopies() {
+    public List<Copy> getAllCopies() {
         return copyRepository.getAllCopies();
+    }
+
+    public List<Copy> getAllAvailable(Long titleId) {
+        return copyRepository.getAllAvailableCopies(titleId);
     }
 
     public Copy getCopy(Long id) throws CopyNotFoundException{
         return copyRepository.getCopy(id).orElseThrow(CopyNotFoundException::new);
     }
 
-    public void deleteCopy(Long id) {
-        copyRepository.deleteById(id);
+    public void deleteCopy(Long id) throws CopyNotFoundException {
+        if (copyRepository.getCopy(id).isPresent()) {
+            copyRepository.deleteById(id);
+        } else throw new CopyNotFoundException();
     }
 
     public void saveCopy(final Copy copy) { copyRepository.save(copy); }
-
-    public Copy updateCopy(final Copy copy) throws CopyNotFoundException {
-        getCopy(copy.getId());
-        return  copyRepository.save(copy);
-    }
 
     public void borrowCopy(final Copy copy) {
         copy.setBorrowed(true);
@@ -45,9 +46,5 @@ public class CopyService {
     public boolean isCopyAvailable(final Long id) throws CopyNotFoundException {
         Copy copy = getCopy(id);
         return !copy.isBorrowed();
-    }
-
-    public List<Copy> getAllAvailable(Long titleId) {
-        return copyRepository.getAllAvailableCopies(titleId);
     }
 }
