+++ src/main/java/com/library/service/ReaderService.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class ReaderService {
     private final ReaderRepository readerRepository;
 
-    public List<Reader> getReaders() {
+    public List<Reader> getAllReaders() {
         return readerRepository.getAllReaders();
     }
 
@@ -21,16 +21,19 @@ public class ReaderService {
         return readerRepository.getReader(id).orElseThrow(ReaderNotFoundException::new);
     }
 
-    public void deleteReader(Long id) {
-        readerRepository.deleteById(id);
+    public void deleteReader(Long id) throws ReaderNotFoundException {
+        if (readerRepository.getReader(id).isPresent()) {
+            readerRepository.deleteById(id);
+        } else throw new ReaderNotFoundException();
+    }
+
+    public Reader updateReader(final Reader reader) throws ReaderNotFoundException {
+        if (readerRepository.getReader(reader.getId()).isEmpty()) throw new ReaderNotFoundException();
+        return readerRepository.save(reader);
     }
 
     public void saveReader(final Reader reader) {
         readerRepository.save(reader);
     }
 
-    public Reader updateReader(final Reader reader) throws ReaderNotFoundException {
-        getReader(reader.getId());
-        return  readerRepository.save(reader);
-    }
 }
